requirements.txt
rasa==3.6.16
rasa-sdk==3.6.2
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.2
python-dotenv==1.0.1
Unidecode==1.3.8